requirements.txt
amqp==5.1.1
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
kombu==5.2.4
packaging==21.3
prompt-toolkit==3.0.30
pyaes==1.6.1
pyparsing==3.0.9
Pyrogram==2.0.43
PySocks==1.7.1
python-dotenv==0.20.0
pytz==2022.2.1
redis==4.3.4
six==1.16.0
TgCrypto==1.2.3
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
